# Conflicts:
#	CHT/plugin/network_service/fermat-cht-plugin-network-service-chat-bitdubai/src/main/java/com/bitdubai/fermat_cht_plugin/layer/network_service/chat/developer/bitdubai/version_1/structure/ChatMetadataRecord.java

+++ BCH/plugin/crypto_vault/fermat-bch-plugin-crypto-vault-assets-over-bitcoin-bitdubai/src/main/java/com/bitdubai/fermat_bch_plugin/layer/asset_vault/developer/bitdubai/version_1/structure/VaultKeyHierarchyMaintainer.java
@@ -159,7 +159,7 @@ class VaultKeyHierarchyMaintainer implements Agent {
                 /**
                  * I will calculate the current threshold to see if we need to create new keys
                  */
-                currentThreshold = (int) Math.round(100 - ((currentUsedKeys * 100) / currentGeneratedKeys));
+                currentThreshold = Math.round(100 - ((currentUsedKeys * 100) / currentGeneratedKeys));
 
                 List<ECKey> keys;
                 if (currentThreshold <= VaultKeyMaintenanceParameters.KEY_PERCENTAGE_GENERATION_THRESHOLD) {
